deblurgan/requirements.txt
albumentations==1.0.3 
cycler==0.10.0 
dataclasses==0.8 
decorator==4.4.2 
fire==0.4.0 
glog==0.3.1 
imageio==2.9.0 
joblib==1.0.1 
kiwisolver==1.3.1 
matplotlib==3.3.4 
munch==2.5.0 
networkx==2.5.1 
numpy==1.19.5 
opencv-python-headless==4.5.3.56 
pandas==1.1.5 
Pillow==8.3.1 
pip==21.2.4 
pkg_resources==0.0.0 
pretrainedmodels==0.7.4 
protobuf==3.17.3 
pthflops==0.4.1 
pyparsing==2.4.7 
pyssim==0.4 
python-dateutil==2.8.2 
python-gflags==3.1.2 
pytz==2021.1 
PyWavelets==1.1.1 
PyYAML==5.4.1 
scikit-image==0.17.2 
scipy==1.5.4 
setuptools==39.0.1 
six==1.16.0 
tensorboardX==2.4 
termcolor==1.1.0 
thop==0.0.31.post2005241907 
tifffile==2020.9.3 
timm==0.4.12 
torch==1.9.0 
torchstat==0.0.7 
torchsummary==1.5.1 
torchvision==0.10.0 
tqdm==4.62.1 
typing-extensions==3.10.0.0 
